requirements.txt
fastapi
uvicorn
pydantic-settings
aiogram>=3.0.0
python-dotenv
pillow
requests
beautifulsoup4
psycopg[binary]
torch
open-clip-torch
mlx-vlm
